% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NP.R
\name{NP}
\alias{NP}
\title{Calculate Nitrogen Productivity (NP)}
\usage{
NP(GR, PlantN)
}
\arguments{
\item{GR}{Plant relative growth rate value}

\item{PlantN}{A numeric vector of values for plant N concentration.}
}
\value{
The calculated NP value.
}
\description{
The ratio of the relative growth rate to the concentration of N in plant tissues.
Calculate Nitrogen Productivity using the formula: NP = GR/PlantN
}
\examples{
GR <- 15
PlantN <- c(12.1, 8.99, 12.89, 13.11)
NP(GR, PlantN)
}
\references{
Primary: Congreves, K. A., Otchere, O., Ferland, D., Farzadfar, S., Williams, S., & Arcand, M. M. (2021, June 4).
Nitrogen Use Efficiency Definitions of Today and Tomorrow. Frontiers in Plant Science, 12.
https://doi.org/10.3389/fpls.2021.637108

Secondary: Berendse, F., and Aerts, R. (1987).
Nitrogen-use-efficiency: a biologically meaningful definition? Funct. Ecol. 1, 293–296.
}
